[3.0] [Img_optm] Fixed large volume image table storage issue. (#328956)

+++ litespeed-cache/src/data_structure/avatar.sql
@@ -1,4 +1,4 @@
-  `id` int(11) unsigned NOT NULL AUTO_INCREMENT,
+  `id` bigint(20) unsigned NOT NULL AUTO_INCREMENT,
   `url` varchar(1000) NOT NULL DEFAULT '',
   `md5` varchar(128) NOT NULL DEFAULT '',
   `dateline` int(11) NOT NULL DEFAULT '0',
